+++ Library/Homebrew/extend/os/readall.rb
@@ -1,4 +1,4 @@
 # typed: strict
 # frozen_string_literal: true
 
-require "extend/os/linux/readall" if OS.linux?
+require "extend/os/mac/readall" if OS.mac?
